ui:adds storgeclass support
This patch updates job submission to include a
storageclass, and has corresponding changes in the
profiles view so you can see what storageclass was
used for a specfific job. This storageclass information
is coming from the /status call that describes the
workers. In prior versions workers was just an int, now
it's an object with attributes corresponding to the
storageclasses of the fioworker pods.

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>